+++ src/chain/pallet/ismp/index.ts
@@ -1,0 +1,26 @@
+export { IsmpModule } from './IsmpModule';
+export type {
+    StateMachineQuery,
+    PostRequest,
+    GetRequest,
+    IsmpRequest,
+    ChannelHeight,
+} from './IsmpModule';
+export type {
+    StateMachineId,
+    StateMachineEnum,
+    RequestDispatchedEvent,
+    IsmpRequestEvent,
+    RequestTimedOutEvent,
+    MessageReceivedEvent,
+    RejectReason,
+    MessageRejectedEvent,
+    GetResponseReceivedEvent,
+    SourceChangedEvent,
+    StateMachineUpdatedEvent,
+    StateCommitmentVetoedEvent,
+    RequestHandledEvent,
+    TimeoutHandledEvent,
+    IsmpMessagingEvent,
+    IsmpEvent,
+} from './events';