+++ src/index.ts
@@ -42,6 +42,7 @@ export * from './chain/rpc/index';
 export * from './chain/pallet/shielded-pool/index';
 export * from './chain/pallet/zk-verifier/index';
 export * from './chain/pallet/relayer/index';
+export * from './chain/pallet/ismp/index';
 export { toTxResult, feePaidFrom, signAndSubmitTx } from './chain/tx';
 export type { UnsafeTx, UnsafeTxOptions, SubmitOptions } from './chain/tx';
 export type { TxFinalizedPayload } from 'polkadot-api';